requirements.txt
pandas
numpy
scikit-learn
xgboost
imbalanced-learn
fastapi
uvicorn
joblib
pytest
hypothesis
pyyaml
